factoring out repeated componen
