+++ services/ui-src/src/components/pages/Dashboard/DashboardTable.tsx
@@ -1,7 +1,6 @@
 // components
 import { Button, Image, Td, Tr } from "@chakra-ui/react";
-import { Table } from "components";
-import { Spinner } from "@cmsgov/design-system";
+import { AdminActionButtons, Table } from "components";
 // utils
 import { AnyObject, ReportMetadataShape } from "types";
 import { convertDateUtcToEt } from "utils";
@@ -122,65 +121,6 @@ const DateFields = ({ report, reportType }: DateFieldProps) => {
 interface DateFieldProps {
   report: ReportMetadataShape;
   reportType: string;
-}
-
-const AdminActionButtons = ({
-  report,
-  reportType,
-  reportId,
-  archiveReport,
-  archiving,
-  locking,
-  lockReport,
-  sxOverride,
-}: AdminActionButtonProps) => {
-  return (
-    <>
-      {reportType === "MLR" && (
-        <Td sx={sxOverride.adminActionCell}>
-          <Button
-            variant="link"
-            sx={sxOverride.adminActionButton}
-            onClick={() => lockReport!(report)}
-          >
-            {locking && reportId === report.id ? (
-              <Spinner size="small" />
-            ) : report?.locked ? (
-              "Unlock"
-            ) : (
-              "Lock"
-            )}
-          </Button>
-        </Td>
-      )}
-      <Td sx={sxOverride.adminActionCell}>
-        <Button
-          variant="link"
-          sx={sxOverride.adminActionButton}
-          onClick={() => archiveReport(report)}
-        >
-          {archiving && reportId === report.id ? (
-            <Spinner size="small" />
-          ) : report?.archived ? (
-            "Unarchive"
-          ) : (
-            "Archive"
-          )}
-        </Button>
-      </Td>
-    </>
-  );
-};
-
-interface AdminActionButtonProps {
-  report: ReportMetadataShape;
-  reportType: string;
-  reportId: string | undefined;
-  archiveReport: Function;
-  archiving: boolean;
-  locking?: boolean;
-  lockReport?: Function;
-  sxOverride: AnyObject;
 }
 
 const sx = {
