+++ services/ui-src/src/components/pages/Dashboard/MobileDashboardTable.tsx
@@ -1,6 +1,6 @@
 // components
 import { Box, Button, Flex, Image, Text } from "@chakra-ui/react";
-import { Spinner } from "@cmsgov/design-system";
+import { AdminActionButtons } from "./AdminActionButtons";
 // utils
 import { AnyObject, ReportMetadataShape } from "types";
 import { convertDateUtcToEt } from "utils";
@@ -127,61 +127,6 @@ const DateFields = ({ report, reportType }: DateFieldProps) => {
 interface DateFieldProps {
   report: ReportMetadataShape;
   reportType: string;
-}
-
-const AdminActionButtons = ({
-  report,
-  reportType,
-  reportId,
-  archiveReport,
-  archiving,
-  locking,
-  lockReport,
-  sxOverride,
-}: AdminActionButtonProps) => {
-  return (
-    <>
-      {reportType === "MLR" && (
-        <Button
-          variant="link"
-          sx={sxOverride.adminActionButton}
-          onClick={() => lockReport!(report)}
-        >
-          {locking && reportId === report.id ? (
-            <Spinner size="small" />
-          ) : report?.locked ? (
-            "Unlock"
-          ) : (
-            "Lock"
-          )}
-        </Button>
-      )}
-      <Button
-        variant="link"
-        sx={sxOverride.adminActionButton}
-        onClick={() => archiveReport(report)}
-      >
-        {archiving && reportId === report.id ? (
-          <Spinner size="small" />
-        ) : report?.archived ? (
-          "Unarchive"
-        ) : (
-          "Archive"
-        )}
-      </Button>
-    </>
-  );
-};
-
-interface AdminActionButtonProps {
-  report: ReportMetadataShape;
-  reportType: string;
-  reportId: string | undefined;
-  archiveReport: Function;
-  archiving: boolean;
-  locking?: boolean;
-  lockReport?: Function;
-  sxOverride: AnyObject;
 }
 
 const sx = {
